+++ src/Curry/Base/LexComb.hs
@@ -20,7 +20,7 @@
 -}
 module Curry.Base.LexComb
   ( -- * Types
-    Symbol (..), Indent, Context, P, MessageM, SuccessP, FailP, Lexer
+    Symbol (..), Indent, Context, P, CYM, SuccessP, FailP, Lexer
 
     -- * Monadic functions
   , parse, applyLexer, returnP, thenP, thenP_, failP, liftP, closeP0, closeP1
@@ -35,7 +35,7 @@ module Curry.Base.LexComb
 
 import Data.Char           (digitToInt)
 
-import Curry.Base.Message  (MessageM, failWithAt)
+import Curry.Base.Monad    (CYM, failMessageAt)
 import Curry.Base.Position (Position, first)
 
 
@@ -58,11 +58,11 @@ type P a = Position   -- ^ Current source code position
         -> Bool       -- ^ Flag whether the beginning of a line should be
                       --   parsed, which requires layout checking
         -> Context    -- ^ context as a stack of 'Indent's
-        -> MessageM a
+        -> CYM a
 
 -- |Apply a lexer on a 'String' to lex the content. The second parameter
 -- requires a 'FilePath' to use in the 'Position'
-parse :: P a -> FilePath -> String -> MessageM a
+parse :: P a -> FilePath -> String -> CYM a
 parse p fn s = p (first fn) s True []
 
 -- ---------------------------------------------------------------------------
@@ -105,7 +105,7 @@ p1 `thenP_` p2 = p1 `thenP` \_ -> p2
 
 -- |Fail to lex on a 'Position', given an error message
 failP :: Position -> String -> P a
-failP pos msg _ _ _ _ = failWithAt pos msg
+failP pos msg _ _ _ _ = failMessageAt pos msg
 
 -- |Apply a pure function to the lexers result
 liftP :: (a -> b) -> P a -> P b
@@ -131,7 +131,7 @@ pushContext col cont pos s bol ctxt = cont pos s bol (col : ctxt)
 -- |Pop an 'Indent' from the context, decreasing the levels of indentation
 popContext :: P a -> P a
 popContext cont pos s bol (_ : ctxt) = cont pos s bol ctxt
-popContext _    pos _ _   []         = failWithAt pos $
+popContext _    pos _ _   []         = failMessageAt pos $
   "Parse error: popping layout from empty context stack. " ++
   "Perhaps you have inserted too many '}'?"